// backend/src/common/constants/enums.ts
export enum RoleEnum {
    ADMIN = 'ADMIN',
    USER = 'USER',
}

export enum JwtEnum {
    ACCESS = 'ACCESS',
    REFRESH = 'REFRESH',
    ACTIVATE = 'ACTIVATE',
}

export enum ResEnum {
    SUCCESS = 'Success 👍',
    FAILURE = 'Failure 👎',
}

export enum StatusOrder {
    CREATED = 'CREATED',
    ACTIVE = 'ACTIVE',
    MODIFIED = 'MODIFIED',
    PROCESSED = 'PROCESSED',
    CLOSED = 'CLOSED',
    ARCHIVED = 'ARCHIVED'
}